Update Broken links and sidebar close onTransition
