+++ public/sitemap.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>http://localhost:1337/portfolio</loc><lastmod>2025-04-04T12:02:07.415Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:1337/blog/category</loc><lastmod>2025-04-04T12:02:07.415Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:1337/blog</loc><lastmod>2025-04-04T12:02:07.415Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:1337/contact</loc><lastmod>2025-04-04T12:02:07.415Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:1337/blog/tag</loc><lastmod>2025-04-04T12:02:07.415Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:1337</loc><lastmod>2025-04-04T12:02:07.415Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:1337/portfolio</loc><lastmod>2025-04-05T04:21:01.506Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:1337/blog/category</loc><lastmod>2025-04-05T04:21:01.507Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:1337/blog/tag</loc><lastmod>2025-04-05T04:21:01.507Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:1337/blog</loc><lastmod>2025-04-05T04:21:01.507Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:1337/contact</loc><lastmod>2025-04-05T04:21:01.507Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:1337</loc><lastmod>2025-04-05T04:21:01.507Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
 <url><loc>http://localhost:1337/portfolio/arete-health</loc><changefreq>monthly</changefreq><priority>0.8</priority></url>
 <url><loc>http://localhost:1337/portfolio/kanarys</loc><changefreq>monthly</changefreq><priority>0.8</priority></url>
 <url><loc>http://localhost:1337/portfolio/raush</loc><changefreq>monthly</changefreq><priority>0.8</priority></url>
